Cleanup and added a first draft of work stealing algorithm.

+++ backend/distributed/sfc/mars_sfc_utils.hpp
@@ -8,6 +8,7 @@
  * @author Daniel Ganellari
  */
 
+#include <_types/_uint64_t.h>
 #include "mars_base.hpp"
 #include "mars_config.hpp"
 #include "mars_globals.hpp"
@@ -19,26 +20,18 @@ struct maxTreeLevel
 {
 };
 
-template<>
-struct maxTreeLevel<unsigned> : integral_constant<unsigned, 10>
-{
-};
+template <>
+struct maxTreeLevel<unsigned> : integral_constant<unsigned, 10> {};
 
-template<>
-struct maxTreeLevel<unsigned long long> : integral_constant<unsigned, 21>
-{
-};
+template <>
+struct maxTreeLevel<uint64_t> : integral_constant<unsigned, 21> {};
 
-template<>
-struct maxTreeLevel<unsigned long> : integral_constant<unsigned, 21>
-{
-};
+template <>
+struct maxTreeLevel<unsigned long> : integral_constant<unsigned, 21> {};
 
 //! @brief maximum integer coordinate
-template<class KeyType>
-struct maxCoord : integral_constant<unsigned, (1u << maxTreeLevel<KeyType>{})>
-{
-};
+template <class KeyType>
+struct maxCoord : integral_constant<unsigned, (1u << maxTreeLevel<KeyType>{})> {};
 
 }  // namespace mars
 #endif
